% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aSAH_Robin}
\alias{aSAH_Robin}
\title{S100beta biomarker concentration and health outcome}
\format{
A tibble with 113 rows and 2 variables:
\describe{
  \item{\code{obs}}{a binary variable indicating positive or negative health outcome.}
  \item{\code{forc}}{S100beta biomarker concentration}
}
}
\source{
'pROC' package, https://cran.r-project.org/package=pROC

Reference:
  Robin X, Turck N, Hainard A, Tiberti N, Lisacek F, Sanchez J-C,
  Müller M (2011). pROC: An open-source package for R and S+ to analyze and
  compare ROC curves. BMC Bioinformatics, 12, 77.
}
\usage{
aSAH_Robin
}
\description{
A modified version of the data set 'aSAH' from the pROC package.
}
\keyword{datasets}
