% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaROC.R
\docType{package}
\name{betaROC-package}
\alias{betaROC}
\alias{betaROC-package}
\title{betaROC: Minimum Distance Estimation for ROC curves}
\description{
Estimating ROC curve models using L2 distance minimization. See
    Gneiting and Vogel (2018) <arXiv:1809.04808>.
}
\author{
\strong{Maintainer}: Alexander I. Jordan \email{alexander.jordan@h-its.org} [contributor]

Authors:
\itemize{
  \item Peter Vogel
}

}
\keyword{internal}
