% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_model.frame}
\alias{check_model.frame}
\title{Checks the correctness of model.frame inputs}
\usage{
check_model.frame(mf)
}
\arguments{
\item{mf}{Model frame}
}
\value{
Returns a model frame
}
\description{
Checks the correctness of model.frame inputs
}
\details{
Ensures that observations are binary and no infinite values are
  present in the forecasts
}
