% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_empROC}
\alias{fill_empROC}
\title{Fill ROC curve by linear interpolation}
\usage{
fill_empROC(empROC, spacing = 0.05)
}
\arguments{
\item{empROC}{\code{data.frame} containing true positive (TPR) and false
positive rates (FPR), ordered by FPR, then TPR}

\item{spacing}{optional; defines approximately the spacing at which
interpolated values are computed}
}
\value{
\code{data.frame} empROC, containing FPR and TPR values, ordered by
  FPR, then TPR
}
\description{
Fill ROC curve by linear interpolation
}
\note{
Check sorting by FPR, then TPR
}
