% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MDE_fit.R
\name{fit_initial_pars}
\alias{fit_initial_pars}
\title{Fits initial parameters for the minimum distance estimation of an empirical
ROC curve}
\usage{
fit_initial_pars(empROC, MDE_info)
}
\arguments{
\item{empROC}{\code{data.frame} with true positive (TPR) and false positive
(FPR) values of the empirical ROC curve}

\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit}
}
\description{
Fits initial parameters for the minimum distance estimation of an empirical
ROC curve
}
