% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_roc.R
\name{get_TPR}
\alias{get_TPR}
\title{Computes the TPR value for given FPR value and model specification}
\usage{
get_TPR(FPR, pars, MDE_info)
}
\arguments{
\item{FPR}{optional; false positive rates for which true positive rates (TPR)
are required}

\item{pars;}{a vector of parameters}

\item{MDE_info;}{a list with information about the MDE fit}
}
\value{
Returns a numeric vector of TPR entries
}
\description{
Computes the TPR value for given FPR value and model specification
}
