% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_roc.R
\name{get_TPR_beta2p}
\alias{get_TPR_beta2p}
\title{Implementations of get_TPR functions for the different MDE methods}
\usage{
get_TPR_beta2p(FPR, pars)
}
\description{
Implementations of get_TPR functions for the different MDE methods
}
