% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{leftalign_caption}
\alias{leftalign_caption}
\title{\code{align_caption} aligns the caption at the left if the \code{cowplot}
package is available}
\usage{
leftalign_caption(p, caption = NULL, subtitle = NULL)
}
\arguments{
\item{p}{A plot created by \code{ggplot}}

\item{caption}{optional; the caption to be added}

\item{subtitle}{optional; the subtitle to be added}
}
\value{
The output is the same plot but with left-aligned caption and a
 subtitle if provided
}
\description{
\code{align_caption} aligns the caption at the left if the \code{cowplot}
package is available
}
