% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linint_empROC}
\alias{linint_empROC}
\title{Linear interpolation of a ROC curve}
\usage{
linint_empROC(FPR, empROC)
}
\arguments{
\item{FPR}{False positive rate for which linear interpolation is required}

\item{empROC}{\code{data.frame} containing true positive (TPR) and false
positive rates (FPR), ordered by FPR, then TPR}
}
\value{
Vector of FPR and corresponding TPR value
}
\description{
Linear interpolation of a ROC curve
}
\note{
Check sorting by FPR, then TPR
}
