% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDE.R
\name{MDE}
\alias{MDE}
\title{Minimum distance fit for empirical ROC curve}
\usage{
MDE(empROC, MDE_info, maxit = 100, pars_init = NULL)
}
\arguments{
\item{empROC}{\code{data.frame} with columns \code{FPR} and \code{TPR},
false positive and true positive values of an empirical ROC curve.}

\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit. Entries are \code{method} and \code{info}.}

\item{maxit}{maximum number of iterations for the numerical optimization,
a \code{control} parameter for \code{\link[stats]{optim}}.}

\item{pars_init}{optional; a vector of initial parameter values for the
numerical optimization.}
}
\value{
Returns a list with initially estimated parameters and corresponding
  L2 distance and the final estimated parameters as well as the associated L2
  distance
}
\description{
Uses minimum distance estimation to fit a parametric ROC curve model to the
empirical ROC curve at hand.
}
\details{
\code{MDE} relies on optimization of the L2 distance to estimate
  the optimal parameters of a parametric ROC curve model.
}
