% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot.roc}
\alias{plot.roc}
\title{S3 method for \code{roc} object}
\usage{
\method{plot}{roc}(x, which, ...)
}
\arguments{
\item{x}{\code{roc} object}

\item{which}{optional; specifies which type of plots shall be created. If not
set, all plots are created. Possible values are histogramm, empirical, fit,
roc_uncertainty, parameter_uncertainty, and partial matching is supported.}

\item{...}{optional; additional arguments to be passed to the single plotting
functions (see details)}
}
\description{
\code{plot.roc} is the S3
}
\details{
TBD
}
