% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot_binormal}
\alias{plot_binormal}
\title{Plot function for binormal ROC curves}
\usage{
plot_binormal(pars, MDE_info, p, lty, lwd, color)
}
\arguments{
\item{pars}{parameters for the ROC family given by \code{MDE_info}}

\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit. Entries are \code{method} and \code{info}.}

\item{p}{ggplot object to add a line to}

\item{lty}{line type}

\item{lwd}{line width}

\item{color}{line color}
}
\description{
Plot function for binormal ROC curves
}
