% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\alias{plot_empirical}
\alias{plot_beta}
\alias{plot_binormal}
\title{Plot functions for empirical and parametric ROC curves}
\usage{
plot_empirical(x, ...)

plot_beta(pars, MDE_info, p, lty, lwd, color, ...)

plot_binormal(pars, MDE_info, p, lty, lwd, color, ...)
}
\arguments{
\item{x}{an object of class \code{'roc'}.}

\item{...}{parameters to be passed to \code{geom_line}.}

\item{pars}{parameters for the ROC family given by \code{MDE_info}}

\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit. Entries are \code{method} and \code{info}.}

\item{p}{ggplot object to add a line to}

\item{lty}{line type}

\item{lwd}{line width}

\item{color}{line color}
}
\description{
Plot functions for empirical and parametric ROC curves
}
