% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot_roc_histogramm}
\alias{plot_roc_histogramm}
\title{Histogramm and support plot for \code{roc} objects}
\usage{
plot_roc_histogramm(x, ...)
}
\arguments{
\item{x}{\code{roc} object}

\item{...}{optional; additional arguments to be passed to the single plotting
functions (see details)}
}
\description{
Histogramm and support plot for \code{roc} objects
}
\details{
to be added
}
