% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_test.R
\name{resample_test}
\alias{resample_test}
\title{Goodness-of-fit test for parametric models of ROC curves}
\usage{
resample_test(roc, MDE_info, m = 100)
}
\arguments{
\item{roc}{an object of class \code{'roc'}.}

\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit. Entries are \code{method} and \code{info}.}

\item{m}{the number of Monte Carlo replications.}
}
\value{
returns a list with entries
  \tabular{ll}{
  \code{L2_empirical} \tab the L2 distance of the fitted ROC curve model to
  the empirical ROC curve \cr
  \code{L2_reference} \tab a vector of L2 distances from the Monte Carlo
  resample procedure \cr
  \code{pval} \tab the p-value calculated as \code{(sum(L2_reference >= L2_empirical) + 1) / (m + 1)}
  }
}
\description{
Goodness-of-fit test for parametric models of ROC curves
}
