% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_test.R
\name{rm_attribute_terms}
\alias{rm_attribute_terms}
\title{Removes the attribute terms from a model.frame}
\usage{
rm_attribute_terms(mf)
}
\arguments{
\item{mf}{Model frame}
}
\value{
Returns a data.frame
}
\description{
Removes the attribute terms from a model.frame
}
