% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_roc.R
\name{roc_empirical}
\alias{roc_empirical}
\title{Computes the empirical ROC curve values}
\usage{
roc_empirical(mf, emp_info)
}
\arguments{
\item{mf}{Model frame as provide by \code{roc}}

\item{emp_info}{Information if unrestricted or concave empirical ROC curve
shall be returned}
}
\value{
Returns a data.frame with true positive rates (TPR) and false
  positive rates (FPR)
}
\description{
Computes the empirical ROC curve values
}
\details{
\code{roc_empirical} computes the true positive (TPR) and
  corresponding false positive rates (FPR) for the empirical data provided in
  \code{mf}. For faster computations, the evaluation utilizes that only at
  the distinct forecast values, changes of FPR or TPR occur.
}
