% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MDE_fit.R
\name{roc_sqe}
\alias{roc_sqe}
\title{Squared error for fit of empirical by parametric ROC curve}
\usage{
roc_sqe(pars, empROC, MDE_info)
}
\arguments{
\item{pars}{Parameter vector corresponding to \code{MDE_info$method}}

\item{empROC}{\code{data.frame} with true positive (TPR) and false positive
(FPR) values of the empirical ROC curve}

\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit}
}
\description{
Squared error for fit of empirical by parametric ROC curve
}
\details{
Computes the sum of the squared errors between an empirical ROC
  curve and a parametric one specified by pars and MDE_info, evaluated at the
  FPR values of the empirical ROC curve. In case a concave binormal ROC curve
  is specified, the parameters have to be adjusted in that sigma = 1 is
  inserted into the pars vector.
}
