% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_test.R
\name{roc_test}
\alias{roc_test}
\title{Tests two objects of class \code{roc}}
\usage{
roc_test(roc1, roc2, testname, parametric, ...)
}
\arguments{
\item{roc1}{Element of class \code{roc}}

\item{roc2}{Element of class \code{roc}; necessary for tests of equal AUC or
equal ROC curves}

\item{testname}{Name of the test to be performed. The choices
goodness-of-fit, auc, and roc are available and partial matching is
supported.}

\item{parametric}{no idea}

\item{...}{beats me}
}
\value{
Returns an object of class \code{roctest}, containing ...
}
\description{
Tests two objects of class \code{roc}
}
\details{
The function allows to test the goodness-of-fit of one ROC curve as
  well as to test for AUC or ROC curve equality.

  The goodness-of-fit tests in case for one provided ROC curves are based on
  Monte Carlo type approaches to estimate the .. and require comparatively
  long computation times.

  The tests for AUC and ROC curve inequality are based on the asymptotic
  distribution in case of parametric beta and binormal ROC curves and, so
  far, only available for the 2-parameter cases.
}
\note{
GOODNESS-OF-FIT - Parametric - Non-parametric (?) AUC value -
  Parametric - Non-parametric (?) - for paired and unpaired data ROC curve -
  Parametric - Non-parametric - for paired and unpaired data
}
