% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MDE_fit.R
\name{shift_ipar_beta}
\alias{shift_ipar_beta}
\title{Shifts beta parameters into the triangular region corresponding to concave
beta ROC curves}
\usage{
shift_ipar_beta(pars, MDE_info, eps = 0.1)
}
\arguments{
\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit}

\item{eps}{optional; specifies how far into the triangular parameter region
corresponding to a concave beta ROC curve the beta parameters are shifted}
}
\description{
Shifts beta parameters into the triangular region corresponding to concave
beta ROC curves
}
\details{
WRITE ABOUT SHIFTING ETC.
}
