% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MDE_fit.R
\name{visualize_ipar_beta}
\alias{visualize_ipar_beta}
\title{Only for test purposes}
\usage{
visualize_ipar_beta(MDE_info, eps)
}
\arguments{
\item{MDE_info}{\code{list} containing information about the required minimum
distance estimation (MDE) fit}

\item{eps}{optional; specifies how far into the triangular parameter region
corresponding to a concave beta ROC curve the beta parameters are shifted}
}
\description{
Only for test purposes
}
\note{
Visualizes the shift in initial pars in case of concave beta fits
}
