% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WS_Vogel}
\alias{WS_Vogel}
\title{Precipitation forcasts and occurence in the West Sahel region}
\format{
A tibble with 5449 rows and 2 variables:
\describe{
  \item{\code{obs}}{a binary variable indicating occurence of precipitation}
  \item{\code{forc}}{probability of precipition}
}
}
\source{
Vogel P, Knippertz P, Fink AH, Schlueter A, Gneiting T (2018). Skill of
  global raw and postprocessed ensemble predictions of rainfall over northern
  tropical Africa. Weather and Forecasting, 33, 369--388.

This data set contains modified historic products from the European Center for
Medium-Range Weather Forecasts (ECMWF, https://www.ecmwf.int/), specifically:
ensemble forecasts of precipitation that have been summarized to a probability of
precipitation, and historical observations for the occurence of precipitation.
The ECMWF licenses the use of expired real-time data products under the
Creative Commons Attribution 4.0 International license
(CC-BY 4.0, https://creativecommons.org/licenses/by/4.0/).
}
\usage{
WS_Vogel
}
\description{
A modified version of the data set used in the study by Vogel et al (2018).
}
\keyword{datasets}
