% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Receiver Operating Characteristic (ROC) object}
\usage{
roc(formula, data, emp_info, fit_method, fit_info)
}
\arguments{
\item{formula}{Formula specifying observation and forecast}

\item{data}{optional; data frame which incorporates the variables named in
\code{formula}. If \code{data} is not specified, \code{roc} searches
for the variables specified in \code{formula} in .GlobalEnv.}

\item{method}{Method specifies the type of ROC curve to be computed. Partial
matching is supported and the following names are permitted
"empirical" generates only the empirical ROC curve
"bin2p" generates the classical 2-parameter binormal ROC model
"bin3p" generates the 3-parameter binormal ROC model
"beta2p" generate the 2-parameter beta ROC model
"beta3p_v" generates the 3-parameter beta ROC model with a
  vertical straight edge at (0,0)
"beta3p_h" generates the 3-parameter beta ROC model with a
  horizontal straight edge at (1,1)
"beta4p" generates the 4-parameter beta ROC model}
}
\value{
The output is an object of class \code{roc} which is a list
  containing the following components:
  \tabular{ll}{
    \code{x} \tab TBD \cr
    \code{y} \tab TBD
  }
}
\description{
\code{roc} constructs and returns an object of class \code{roc}.
}
\details{
TBD
}
